* feature/ec-unit:
  Added EC conversion functionality
  Changed ppm type to double
  Added EC setting and re-grouped other settings

+++ app/src/main/java/me/anon/model/Water.java
@@ -18,7 +18,7 @@ import lombok.experimental.Accessors;
 @Accessors(prefix = {"m", ""}, chain = true)
 public class Water extends Action
 {
-	private Long ppm;
+	private Double ppm;
 	private Double ph;
 	private Double runoff;
 	private Double amount;
